logiciel-gestion/desktop_app/requirements.txt
PySide6>=6.5
SQLAlchemy>=1.4
psycopg2-binary>=2.9
bcrypt>=4.0
python-dotenv>=1.0
